Updates to #110 removed more references to mongodb
