+++ Source/Tests/app.config
@@ -1,7 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
   <connectionStrings>
-    <add name="MongoConnectionString" connectionString="mongodb://localhost/exceptionless-test" />
     <add name="RedisConnectionString" connectionString="" />
     <add name="ElasticSearchConnectionString" connectionString="http://localhost:9200" />
     <add name="AzureStorageConnectionString" connectionString="UseDevelopmentStorage=true;" />
